Add files via upload
feat: compare cards

Add function to compare two cards. Setting the array's length to zero is the preferred method to empty the array containing the list of open cards as it has good performance and little complications compared to other methods such as assigning an empty array or using a loop to pop each item. Splice was not chosen as it returns a copy of the array as well